+++ src/mesh/meshblock.hpp
@@ -171,7 +171,8 @@ class MeshBlock : public std::enable_shared_from_this<MeshBlock> {
     Kokkos::deep_copy(exec_space, dst, src);
   }
 
-  void AllocateSparse(std::string const &label);
+  void AllocateSparse(std::string const &label, bool only_control = false,
+                      bool flag_uninitialized = false);
 
   void AllocSparseID(std::string const &base_name, const int sparse_id) {
     AllocateSparse(MakeVarLabel(base_name, sparse_id));
@@ -197,6 +198,12 @@ class MeshBlock : public std::enable_shared_from_this<MeshBlock> {
     return true;
   }
 #endif
+
+  void SetAllVariablesToInitialized() {
+    auto &stages = meshblock_data.Stages();
+    std::for_each(stages.begin(), stages.end(),
+                  [](auto &pair) { pair.second->SetAllVariablesToInitialized(); });
+  }
 
   template <class... Args>
   inline void par_for(Args &&...args) {